fix : 자유게시판(Forum) 댓글 삭제로직 버그 fix (댓글 (대댓글의 대장) 이 삭제되지 않는 버그 발생 -> query에서 대댓글을 가져올때 댓글(parentId==null, groupCommentId==null)을 가져오지 않아 최하단 대댓글 삭제시 대댓글의 루트를 삭제할 수 있음에도 쿼리에서 id값을 가져오지 못해 접근하지 않아 삭제하지 못하는 현상 fix

+++ src/main/java/dgu/edu/dnaapi/service/Forums/ForumCommentsService.java
@@ -169,13 +169,11 @@ public class ForumCommentsService {
     private void shouldParentCommentBeDeleted(Map<Long, ForumCommentVO> forumCommentsMap, Long parentCommentId){
         if(forumCommentsMap.containsKey(parentCommentId)){
             ForumCommentVO parentComment = forumCommentsMap.get(parentCommentId);
-            if(checkAllDeleted(parentComment.getChildrenComment())){
-                parentComment.setDeletedStatus();
+            if(parentComment.isDeleted() && checkAllDeleted(parentComment.getChildrenComment())){
                 forumCommentsRepository.deleteById(parentCommentId);
                 if(parentComment.hasParentComment())
-                    shouldParentCommentBeDeleted(forumCommentsMap, forumCommentsMap.get(parentCommentId).getParentCommentId());
+                    shouldParentCommentBeDeleted(forumCommentsMap, parentComment.getParentCommentId());
             }
         }
-
     }
 }
